[TS-276] Refactor: 컴포넌트, 함수 파일 분리 후 이동
- Star 컴포넌트 파일로 분리 후 StarMain 폴더 내부로 이동
- 별 생성하는 함수 파일로 분리 후 util로 이동
